requirements.txt
anyio==3.6.1
appnope==0.1.3
asttokens==2.0.8
asyncio==3.4.3
backcall==0.2.0
bidict==0.22.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
eventlet==0.33.1
executing==1.1.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
fonttools==4.37.4
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
ipykernel==6.16.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mysql-connector-python==8.0.29
nest-asyncio==1.5.6
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
polygon==1.0.9
polygon-api-client==1.1.0
prompt-toolkit==3.0.31
protobuf==4.21.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.1
pyzmq==24.0.1
rel==0.4.7
requests==2.28.1
rfc3986==1.5.0
seaborn==0.12.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
stack-data==0.5.1
tornado==6.2
traitlets==5.4.0
urllib3==1.26.10
wcwidth==0.2.5
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.2.2
